database created

+++ db/migrations/20180514183252_initial.js
@@ -3,16 +3,16 @@ exports.up = function(knex, Promise) {
     knex.schema.createTable('maps', function(table) {
       table.increments('id').primary();
       table.string('region');
-      table.integer('center_lat');
-      table.integer('center_long');
+      table.decimal('center_lat');
+      table.decimal('center_long');
       table.timestamps(true, true);
     }),
 
     knex.schema.createTable('pins', function(table) {
       table.increments('id').primary();
       table.string('name');
-      table.integer('latitude');
-      table.integer('longitude');
+      table.decimal('latitude');
+      table.decimal('longitude');
       table.integer('map_id');
       table.foreign('map_id').references('maps.id');
       table.timestamps(true, true);
